moving is_schema_valid and linting
