+++ checkov/github_actions/runner.py
@@ -1,16 +1,19 @@
 from __future__ import annotations
 
+import json
 import os
+import logging
 from collections.abc import Iterable
 from typing import TYPE_CHECKING, Any, cast
+
+from schema import SchemaError  # type: ignore
 
 from checkov.common.images.image_referencer import ImageReferencer, Image
 from checkov.common.bridgecrew.check_type import CheckType
 from checkov.common.util.consts import START_LINE, END_LINE
 from checkov.github_actions.checks.registry import registry
-from checkov.github_actions.schema_validator import is_schema_valid
+from checkov.github_actions.schema_validator import schema
 from checkov.yaml_doc.runner import Runner as YamlRunner
-
 if TYPE_CHECKING:
     from checkov.common.checks.base_check_registry import BaseCheckRegistry
 
@@ -173,3 +176,17 @@ class Runner(YamlRunner, ImageReferencer):
                 return cast(str, name)
 
         return ""
+
+    @staticmethod
+    def is_schema_valid(config: dict[str, Any] | list[dict[str, Any]]) -> bool:
+        valid = False
+        try:
+            schema.validate(config)
+            valid = True
+        except SchemaError as e:
+            logging.info(f'Given entity configuration does not match the schema\n'
+                         f'config={json.dumps(config, indent=4)}\n'
+                         f'schema={json.dumps(schema.json_schema("https://example.com/my-schema.json"), indent=4)}')
+            logging.info(f'Error: {e}', exc_info=e)
+
+        return valid
